+++ src/net.cpp
@@ -475,7 +475,6 @@ CNode *CConnman::ConnectNode(CAddress addrConnect, const char *pszDest,
     }
 
     // Connect
-    bool connected = false;
     std::unique_ptr<Sock> sock;
     Proxy proxy;
     CAddress addr_bind;
@@ -488,6 +487,7 @@ CNode *CConnman::ConnectNode(CAddress addrConnect, const char *pszDest,
 
         if (addrConnect.GetNetwork() == NET_I2P && use_proxy) {
             i2p::Connection conn;
+            bool connected{false};
 
             if (m_i2p_sam_session) {
                 connected = m_i2p_sam_session->Connect(addrConnect, conn,
@@ -497,7 +497,7 @@ CNode *CConnman::ConnectNode(CAddress addrConnect, const char *pszDest,
                     LOCK(m_unused_i2p_sessions_mutex);
                     if (m_unused_i2p_sessions.empty()) {
                         i2p_transient_session =
-                            std::make_unique<i2p::sam::Session>(proxy.proxy,
+                            std::make_unique<i2p::sam::Session>(proxy,
                                                                 &interruptNet);
                     } else {
                         i2p_transient_session.swap(
@@ -522,22 +522,17 @@ CNode *CConnman::ConnectNode(CAddress addrConnect, const char *pszDest,
                 addr_bind = CAddress{conn.me, NODE_NONE};
             }
         } else if (use_proxy) {
-            sock = CreateSock(proxy.proxy);
-            if (!sock) {
-                return nullptr;
-            }
-            connected = ConnectThroughProxy(
-                proxy, addrConnect.ToStringAddr(), addrConnect.GetPort(), *sock,
-                nConnectTimeout, proxyConnectionFailed);
+            LogPrintLevel(BCLog::PROXY, BCLog::Level::Debug,
+                          "Using proxy: %s to connect to %s:%s\n",
+                          proxy.ToString(), addrConnect.ToStringAddr(),
+                          addrConnect.GetPort());
+            sock = ConnectThroughProxy(proxy, addrConnect.ToStringAddr(),
+                                       addrConnect.GetPort(),
+                                       proxyConnectionFailed);
         } else {
             // no proxy needed (none set for target network)
-            sock = CreateSock(addrConnect);
-            if (!sock) {
-                return nullptr;
-            }
-            connected =
-                ConnectSocketDirectly(addrConnect, *sock, nConnectTimeout,
-                                      conn_type == ConnectionType::MANUAL);
+            sock = ConnectDirectly(addrConnect,
+                                   conn_type == ConnectionType::MANUAL);
         }
         if (!proxyConnectionFailed) {
             // If a connection to the node was attempted, and failure (if any)
@@ -546,18 +541,13 @@ CNode *CConnman::ConnectNode(CAddress addrConnect, const char *pszDest,
             addrman.Attempt(addrConnect, fCountFailure);
         }
     } else if (pszDest && GetNameProxy(proxy)) {
-        sock = CreateSock(proxy.proxy);
-        if (!sock) {
-            return nullptr;
-        }
         std::string host;
         uint16_t port{default_port};
         SplitHostPort(std::string(pszDest), port, host);
         bool proxyConnectionFailed;
-        connected = ConnectThroughProxy(proxy, host, port, *sock,
-                                        nConnectTimeout, proxyConnectionFailed);
+        sock = ConnectThroughProxy(proxy, host, port, proxyConnectionFailed);
     }
-    if (!connected) {
+    if (!sock) {
         return nullptr;
     }
 
@@ -2276,7 +2266,7 @@ bool CConnman::BindListenPort(const CService &addrBind, bilingual_str &strError,
         return false;
     }
 
-    std::unique_ptr<Sock> sock = CreateSock(addrBind);
+    std::unique_ptr<Sock> sock = CreateSock(addrBind.GetSAFamily());
     if (!sock) {
         strError =
             strprintf(Untranslated("Couldn't open socket for incoming "
@@ -2514,8 +2504,7 @@ bool CConnman::Start(CScheduler &scheduler, const Options &connOptions) {
     Proxy i2p_sam;
     if (GetProxy(NET_I2P, i2p_sam) && connOptions.m_i2p_accept_incoming) {
         m_i2p_sam_session = std::make_unique<i2p::sam::Session>(
-            gArgs.GetDataDirNet() / "i2p_private_key", i2p_sam.proxy,
-            &interruptNet);
+            gArgs.GetDataDirNet() / "i2p_private_key", i2p_sam, &interruptNet);
     }
 
     for (const auto &strDest : connOptions.vSeedNodes) {