+++ test/functional/feature_proxy.py
@@ -15,6 +15,7 @@ Test plan:
     - support no authentication (other proxy)
     - support no authentication + user/pass authentication (Tor)
     - proxy on IPv6
+    - proxy over unix domain sockets
 
 - Create various proxies (as threads)
 - Create nodes that connect to them
@@ -35,9 +36,11 @@ addnode connect to generic DNS name
 - Test passing unknown -onlynet
 """
 
+import os
 import socket
+import tempfile
 
-from test_framework.netutil import test_ipv6_local
+from test_framework.netutil import test_ipv6_local, test_unix_socket
 from test_framework.socks5 import (
     AddressType,
     Socks5Command,
@@ -58,14 +61,22 @@ NET_I2P = "i2p"
 # src/rpc/net.cpp::GetNetworksInfo()
 NETWORKS = frozenset({NET_IPV4, NET_IPV6, NET_ONION, NET_I2P})
 
+# Use the shortest temp path possible since UNIX sockets may have as little as 92-char limit
+socket_path = tempfile.NamedTemporaryFile().name
+
+# TODO: bump these when backporting core#23077
+PROXY_UNIX_IDX = 4
+ONION_UNIX_IDX = 5
+
 
 class ProxyTest(BitcoinTestFramework):
     def set_test_params(self):
-        self.num_nodes = 4
+        self.num_nodes = 6
         self.setup_clean_chain = True
 
     def setup_nodes(self):
         self.have_ipv6 = test_ipv6_local()
+        self.have_unix_sockets = test_unix_socket()
         # Create two proxies on different ports
         # ... one unauthenticated
         self.conf1 = Socks5Configuration()
@@ -87,6 +98,15 @@ class ProxyTest(BitcoinTestFramework):
         else:
             self.log.warning("Testing without local IPv6 support")
 
+        if self.have_unix_sockets:
+            self.conf4 = Socks5Configuration()
+            self.conf4.af = socket.AF_UNIX
+            self.conf4.addr = socket_path
+            self.conf4.unauth = True
+            self.conf4.auth = True
+        else:
+            self.log.warning("Testing without local unix domain sockets support")
+
         self.serv1 = Socks5Server(self.conf1)
         self.serv1.start()
         self.serv2 = Socks5Server(self.conf2)
@@ -94,6 +114,9 @@ class ProxyTest(BitcoinTestFramework):
         if self.have_ipv6:
             self.serv3 = Socks5Server(self.conf3)
             self.serv3.start()
+        if self.have_unix_sockets:
+            self.serv4 = Socks5Server(self.conf4)
+            self.serv4.start()
 
         # We will not try to connect to this.
         self.i2p_sam = ("127.0.0.1", 7656)
@@ -115,6 +138,8 @@ class ProxyTest(BitcoinTestFramework):
             ],
             ["-listen", f"-proxy={ip2}:{port2}", "-proxyrandomize=1"],
             [],
+            [],
+            [],
         ]
         if self.have_ipv6:
             args[3] = [
@@ -123,6 +148,9 @@ class ProxyTest(BitcoinTestFramework):
                 "-proxyrandomize=0",
                 "-noonion",
             ]
+        if self.have_unix_sockets:
+            args[PROXY_UNIX_IDX] = ["-listen", f"-proxy=unix:{socket_path}"]
+            args[ONION_UNIX_IDX] = ["-listen", f"-onion=unix:{socket_path}"]
         self.add_nodes(self.num_nodes, extra_args=args)
         self.start_nodes()
 
@@ -135,7 +163,7 @@ class ProxyTest(BitcoinTestFramework):
         rv = []
         addr = "15.61.23.23:1234"
         self.log.debug(
-            f"Test: outgoing IPv4 connection through node for address {addr}"
+            f"Test: outgoing IPv4 connection through node {node.index} for address {addr}"
         )
         node.addnode(addr, "onetry")
         cmd = proxies[0].queue.get()
@@ -154,7 +182,7 @@ class ProxyTest(BitcoinTestFramework):
         if self.have_ipv6:
             addr = "[1233:3432:2434:2343:3234:2345:6546:4534]:5443"
             self.log.debug(
-                f"Test: outgoing IPv6 connection through node for address {addr}"
+                f"Test: outgoing IPv6 connection through node {node.index} for address {addr}"
             )
             node.addnode(addr, "onetry")
             cmd = proxies[1].queue.get()
@@ -173,7 +201,7 @@ class ProxyTest(BitcoinTestFramework):
         if test_onion:
             addr = "bitcoinostk4e4re.onion:8333"
             self.log.debug(
-                f"Test: outgoing onion connection through node for address {addr}"
+                f"Test: outgoing onion connection through node {node.index} for address {addr}"
             )
             node.addnode(addr, "onetry")
             cmd = proxies[2].queue.get()
@@ -189,8 +217,9 @@ class ProxyTest(BitcoinTestFramework):
 
         addr = "node.noumenon:8333"
         self.log.debug(
-            f"Test: outgoing DNS name connection through node for address {addr}"
+            f"Test: outgoing DNS name connection through node {node.index} for address {addr}"
         )
+
         node.addnode(addr, "onetry")
         cmd = proxies[3].queue.get()
         assert isinstance(cmd, Socks5Command)
@@ -232,6 +261,15 @@ class ProxyTest(BitcoinTestFramework):
                 [self.serv3, self.serv3, self.serv3, self.serv3],
                 False,
                 False,
+            )
+
+        if self.have_unix_sockets:
+            # TODO: when backporting core#23077 test_cjdns=False
+            self.node_test(
+                self.nodes[PROXY_UNIX_IDX],
+                proxies=[self.serv4, self.serv4, self.serv4, self.serv4],
+                auth=True,
+                test_onion=True,
             )
 
         def networks_dict(d):
@@ -306,6 +344,39 @@ class ProxyTest(BitcoinTestFramework):
             assert_equal(n3["onion"]["reachable"], False)
             assert_equal(n3["i2p"]["reachable"], False)
 
+        if self.have_unix_sockets:
+            n5 = networks_dict(nodes_network_info[PROXY_UNIX_IDX])
+            assert_equal(NETWORKS, n5.keys())
+            for net in NETWORKS:
+                if net == NET_I2P:
+                    expected_proxy = ""
+                    expected_randomize = False
+                else:
+                    expected_proxy = "unix:" + self.conf4.addr  # no port number
+                    expected_randomize = True
+                assert_equal(n5[net]["proxy"], expected_proxy)
+                assert_equal(n5[net]["proxy_randomize_credentials"], expected_randomize)
+            assert_equal(n5["onion"]["reachable"], True)
+            assert_equal(n5["i2p"]["reachable"], False)
+            # TODO: core#23077
+            # assert_equal(n5["cjdns"]["reachable"], False)
+
+            n6 = networks_dict(nodes_network_info[ONION_UNIX_IDX])
+            assert_equal(NETWORKS, n6.keys())
+            for net in NETWORKS:
+                if net != NET_ONION:
+                    expected_proxy = ""
+                    expected_randomize = False
+                else:
+                    expected_proxy = "unix:" + self.conf4.addr  # no port number
+                    expected_randomize = True
+                assert_equal(n6[net]["proxy"], expected_proxy)
+                assert_equal(n6[net]["proxy_randomize_credentials"], expected_randomize)
+            assert_equal(n6["onion"]["reachable"], True)
+            assert_equal(n6["i2p"]["reachable"], False)
+            # TODO: core#23077
+            # assert_equal(n6["cjdns"]["reachable"], False)
+
         self.stop_node(1)
 
         self.log.info("Test passing invalid -proxy hostname raises expected init error")
@@ -373,6 +444,19 @@ class ProxyTest(BitcoinTestFramework):
         self.nodes[1].extra_args = ["-onlynet=abc"]
         msg = "Error: Unknown network specified in -onlynet: 'abc'"
         self.nodes[1].assert_start_raises_init_error(expected_msg=msg)
+
+        self.log.info("Test passing too-long unix path to -proxy raises init error")
+        self.nodes[1].extra_args = [f"-proxy=unix:{'x' * 1000}"]
+        if self.have_unix_sockets:
+            msg = f"Error: Invalid -proxy address or hostname: 'unix:{'x' * 1000}'"
+        else:
+            # If unix sockets are not supported, the file path is incorrectly interpreted as host:port
+            msg = f"Error: Invalid port specified in -proxy: 'unix:{'x' * 1000}'"
+        self.nodes[1].assert_start_raises_init_error(expected_msg=msg)
+
+        # Cleanup socket path we established outside the individual test directory.
+        if self.have_unix_sockets:
+            os.unlink(socket_path)
 
 
 if __name__ == "__main__":